rust: update rust toolchain to 1.66.0 and cbindgen to 0.63.0
cbindgen made a breaking change renaming whitelist to allowlist.

The rest is new Clippy warnings.

Clippy needless_borrow is disabled because the new cases it reported
seemed to cause a move/copy. I am not sure if the compiler would have
created the same output in either case. Might revisit this and
re-enable this linter in the future.

+++ src/rust/bitbox02-rust-c/src/async_usb.rs
@@ -36,7 +36,7 @@ pub enum UsbResponse {
 #[no_mangle]
 pub unsafe extern "C" fn rust_async_usb_copy_response(out: *mut bitbox02::buffer_t) -> UsbResponse {
     use bitbox02_rust::async_usb::{copy_response, CopyResponseErr};
-    let dst = core::slice::from_raw_parts_mut((*out).data, (*out).max_len as usize);
+    let dst = core::slice::from_raw_parts_mut((*out).data, (*out).max_len);
     match copy_response(dst) {
         Ok(len) => {
             (*out).len = len as _;
